requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Werkzeug==2.3.7
pandas==2.1.0
numpy==1.25.2
scikit-learn==1.3.0
google-generativeai==0.3.1
python-dotenv==1.0.0
gunicorn==21.2.0
